+++ modules/apps/analytics/analytics-settings-web/src/main/java/com/liferay/analytics/settings/web/internal/constants/AnalyticsSettingsWebKeys.java
@@ -22,7 +22,6 @@ public class AnalyticsSettingsWebKeys {
 	public static final String ANALYTICS_CONFIGURATION =
 		"ANALYTICS_CONFIGURATION";
 
-	public static final String ANALYTICS_USERS_MANAGER =
-		"ANALYTICS_USERS_MANAGER";
+	public static final String ANALYTICS_USERS_UTIL = "ANALYTICS_USERS_UTIL";
 
 }
