+++ modules/apps/analytics/analytics-settings-web/src/main/java/com/liferay/analytics/settings/web/internal/portal/settings/configuration/admin/display/BaseAnalyticsConfigurationScreen.java
@@ -16,7 +16,7 @@ package com.liferay.analytics.settings.web.internal.portal.settings.configuratio
 
 import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
 import com.liferay.analytics.settings.web.internal.constants.AnalyticsSettingsWebKeys;
-import com.liferay.analytics.settings.web.internal.user.AnalyticsUsersManager;
+import com.liferay.analytics.settings.web.internal.util.AnalyticsUsersUtil;
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
@@ -97,12 +97,11 @@ public abstract class BaseAnalyticsConfigurationScreen
 				AnalyticsConfiguration.class, themeDisplay.getCompanyId()));
 
 		httpServletRequest.setAttribute(
-			AnalyticsSettingsWebKeys.ANALYTICS_USERS_MANAGER,
-			analyticsUsersManager);
+			AnalyticsSettingsWebKeys.ANALYTICS_USERS_UTIL, analyticsUsersUtil);
 	}
 
 	@Reference
-	protected AnalyticsUsersManager analyticsUsersManager;
+	protected AnalyticsUsersUtil analyticsUsersUtil;
 
 	@Reference
 	protected ConfigurationProvider configurationProvider;
