+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/init.jsp
@@ -34,8 +34,8 @@ page import="com.liferay.analytics.settings.web.internal.display.context.Organiz
 page import="com.liferay.analytics.settings.web.internal.display.context.OrganizationManagementToolbarDisplayContext" %><%@
 page import="com.liferay.analytics.settings.web.internal.display.context.UserGroupDisplayContext" %><%@
 page import="com.liferay.analytics.settings.web.internal.display.context.UserGroupManagementToolbarDisplayContext" %><%@
-page import="com.liferay.analytics.settings.web.internal.user.AnalyticsUsersManager" %><%@
 page import="com.liferay.analytics.settings.web.internal.util.AnalyticsSettingsUtil" %><%@
+page import="com.liferay.analytics.settings.web.internal.util.AnalyticsUsersUtil" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.ArrayUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
